It was not possible to implicitly use all input fields and specify types for only a subset.
If <fields> was not set and <types> was, then <types> specified the fields to be output.  There
was no way of specifying types for a subset of the output fields without having to explicitly
specify all of the output fields. Now, <fields> may be set to the string 'all' to achieve this.

This retains the ease of use of only specifying <types> to specify output fields and their types.