API breakage.  dbi encoder no longer flushes record queue upon object destruction.
Flushing the queue might cause an exception if a record cannot be
written.  During destruction exceptions are essentially ignored, so
no notification is ever given.  To prevent this, the queue is no longer flushed
during destruction; instead a warning is issued if the queue is not empty.

Explicitly calling <close> will flush the queue.  It will raise an
exception if there is an error.  The <flush> method is also available.