always generate types
Previously type generation (either from data or from default_type) was
supposedly conditional. it really wasn't, it was pretty much always done, and when it was
turned off by encoders it could lead to a bug in the API (e.g. type_index would be incorrect).

So, now, types are always generated.