use Module::Version to check Cpanel::JSON::XS's version rather than loading it with eval
The previous code to check Cpanel::JSON::XS's version used an eval to
load it and the UNIVERSAL::VERSION method.

Unfortunately, loading Cpanel::JSON::XS also loads its version of JSON::PP::Boolean,
which hangs around even if we don't use Cpanel::JSON::XS for the encoding.

One workaround is to use Symbol::delete_package to get rid of the bad
JSON::PP::Boolean, but Module::Version gives access to the version number
without needing to load the module, which is a much better approach.